Moved test together
